App: extracting logging, added CLI
