+++ package.json
@@ -13,7 +13,6 @@
     "body-parser": "^1.17.2",
     "debug": "^2.6.8",
     "express": "^4.15.3",
-    "express-validator": "^3.2.1",
     "getlorem": "^1.0.1"
   },
   "devDependencies": {
